+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/api/BetController.kt
@@ -14,12 +14,12 @@ import org.springframework.web.bind.annotation.*
 @RequestMapping("/api/v1")
 class BetController(@Autowired private val betService: BetService) {
     @GetMapping("/account/{accountId}/bets", produces = ["application/json"])
-    fun handleGetWithAccountId(@PathVariable("accountId") accountId: Int): ResponseEntity<BetListDto> {
+    fun handleGetWithAccountId(@PathVariable("accountId") accountId: Long): ResponseEntity<BetListDto> {
         return ResponseEntity<BetListDto>(betService.getBetsByAccountId(accountId), HttpStatus.OK)
     }
 
     @PostMapping("/account/{accountId}/bets", consumes = ["application/json"], produces = ["application/json"])
-    fun handlePost(@PathVariable("accountId") accountId: Int, @RequestBody betDto: BetDto):
+    fun handlePost(@PathVariable("accountId") accountId: Long, @RequestBody betDto: BetDto):
             ResponseEntity<HttpHeaders> {
         val betId = betService.placeBet(accountId, betDto)
