+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/repository/RollRepository.kt
@@ -3,5 +3,5 @@ package com.betting.simplebettingapi.repository
 import com.betting.simplebettingapi.model.RollModel
 import org.springframework.data.repository.CrudRepository
 
-interface RollRepository : CrudRepository<RollModel, Int> {
+interface RollRepository : CrudRepository<RollModel, Long> {
 }
