+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/repository/TransactionRepository.kt
@@ -4,6 +4,6 @@ import com.betting.simplebettingapi.model.TransactionModel
 import com.betting.simplebettingapi.model.WalletModel
 import org.springframework.data.repository.CrudRepository
 
-interface TransactionRepository : CrudRepository<TransactionModel, Int> {
+interface TransactionRepository : CrudRepository<TransactionModel, Long> {
     fun findAllByWalletOrderByTransactionDtDesc(wallet: WalletModel): List<TransactionModel>
 }
