+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/repository/WalletRepository.kt
@@ -4,8 +4,8 @@ import com.betting.simplebettingapi.model.WalletModel
 import org.springframework.data.repository.CrudRepository
 import org.springframework.data.domain.Pageable;
 
-interface WalletRepository : CrudRepository<WalletModel, Int> {
-    fun findByAccountId(accountId: Int): WalletModel
+interface WalletRepository : CrudRepository<WalletModel, Long> {
+    fun findByAccountId(accountId: Long): WalletModel
 
     fun findAllByOrderByBalanceDesc(pageable: Pageable): List<WalletModel>
 }
